Add PID-Autotune This is the first stage of it. If your ATmega is very small you can temporarly disable stuff like lookahead. It is a little bit bigger than PID, but you only need this when tuning.
To use it enable PID_AUTOTUNE in your printer.x.h.
Now the standard M104 is used to start the autotune.

If you normally print PLA and this at 200° just start it with M104 S200.
Wait for approx. 5 to 10 minutes. When it is ready, the hotend is turned off and you can see the temperature falling.
Now take the last numbers of Ku and Tu. With them you can calculate the PID-values.

Now you could take this two numbers into config-tool to setup your heater. So you don't need an eeprom for that anymore.
Configtool thing are currently not done.