Speed up ConfigTool startup with wx-tricks Fixes #235
Recently ConfigTool has been very slow for me on Ubuntu Linux.
When I run the app there is a 15 second wait before the window is
first displayed.  I bisected the problem and found it was tied to
the number of pins in `pinNames`, and ultimately that it was caused
by a slow initializer in wx.Choice() when the choices are loaded when
the widget is created.  For some reason, moving the load after the
widget is created is significantly faster.  This change reduces my
startup time to just under 4 seconds.

Further speedup could be had by using lazy initialization of the controls.
But the controls are too bound up in the loaded data to make this simple.
Maybe I will attack it later.

There is still a significant delay when closing the window, but I haven't
tracked what causes it.  Maybe it is caused just by destroying all these
pin controls.

In the process of making this change, I wanted to simplify the number of
locations that bothered to copy the pinNames list and, to support lazy
loading, to try to keep the same list in all pinChoice controls.  I noticed
that all the pinChoice controls already have the same parameters passed
to the addPinChoice function which makes them redundant and confusing.
I removed the extra initializers and just rely on pinNames as the only
list option in addPinChoice for now.  Maybe this flexibility is needed
for some reason later, but I can't see a purpose for it now.